+++ src/frontend/component/tokens/Button/Button.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Button = (props) => {
+    return (
+      <button
+        className="Button"
+        onClick={props.action}>
+        {props.description}
+      </button>
+    )
+}
+
+export default Button
